[SPARK-41246][CORE] Detect RDD id overflow instead of silently wrapping
Use AtomicLong for RDD id allocation so crossing Int.MaxValue is visible
instead of wrapping to negative Ints that break BlockId parsing and other
id-keyed state. Keep RDD.id as Int for compatibility with existing callers
such as `int id = rdd.id()`, and add RDD.idLong as the 64-bit identity.

Default spark.rdd.id.overflow.policy=fail aborts further RDD creation with
a clear error. policy=legacy restores wrap-around for emergencies and
BlockId.apply accepts optional minus signs so UnrecognizedBlockId does not
occur for those names. Full end-to-end Long ids in BlockManager/UI remain
follow-up work once the public Int API can be widened.

+++ core/src/test/scala/org/apache/spark/rdd/PipedRDDSuite.scala
@@ -275,7 +275,7 @@ class PipedRDDSuite extends SparkFunSuite with SharedSparkContext with Eventuall
   def generateFakeHadoopPartition(): HadoopPartition = {
     val split = new FileSplit(new Path("/some/path"), 0, 1,
       Array[String]("loc1", "loc2", "loc3", "loc4", "loc5"))
-    new HadoopPartition(sc.newRddId(), 1, split)
+    new HadoopPartition(sc.newRddId().toInt, 1, split)
   }
 
 }
